fix: don't ignore empty dirs when unpacking model and bento
